fix type mismatch for the loss

+++ components/loss.py
@@ -2,4 +2,4 @@ import torch
 
 def basic_loss(y_pred,y_true):
     criterion = torch.nn.CrossEntropyLoss()
-    return criterion(y_pred, y_true)
+    return criterion(y_pred, y_true.long())
